items_users

+++ server/app/Http/Controllers/ItemController.php
@@ -4,9 +4,20 @@ namespace App\Http\Controllers;
 
 use App\Models\Item;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
-class ItemController extends BaseController
+class ItemController extends BaseController implements HasMiddleware
 {
+
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
+    }
+
     public function index()
     {
         $items = Item::all();
@@ -16,6 +27,7 @@ class ItemController extends BaseController
 
     public function store(Request $request)
     {
+
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'required',
@@ -23,8 +35,9 @@ class ItemController extends BaseController
             'price' => 'required|numeric',
             'duration' => 'required|integer',
             'status' => 'required|in:available,rented,unavailable',
-            'lender_id' => 'required|exists:users,id'
         ]);
+
+        $validated['leander_id']  = $request->user->id;
 
         if ($request->hasFile('item_image')) {
             $imagePath = $request->file('item_image')->store('images', 'public');
@@ -36,13 +49,22 @@ class ItemController extends BaseController
         return response()->json(['message' => 'Item created successfully', 'item' => $item], 201);
     }
 
-    public function show(Item $item)
+
+    public function show($id)
     {
-        return response()->json($item, 200);
+        $item = Item::find($id);
+
+        if (!$item) {
+            return $this->sendError('Item not found', 404);
+        }
+        return  $this->sendResponse($item, 200);
     }
+
 
     public function update(Request $request, Item $item)
     {
+        Gate::authorize('modify', $item);
+
         $validated = $request->validate([
             'name' => 'required|string|max:255',
             'description' => 'required',
@@ -59,13 +81,15 @@ class ItemController extends BaseController
 
         $item->update($validated);
 
-        return response()->json(['message' => 'Item updated successfully', 'item' => $item], 200);
+        return $this->sendResponse(['message' => 'Item updated successfully', 'item' => $item], 200);
     }
 
     public function destroy(Item $item)
     {
+        Gate::authorize('modify', $item);
+
         $item->delete();
 
-        return response()->json(['message' => 'Item deleted successfully'], 200);
+        return $this->sendResponse(['message' => 'Item deleted successfully'], 200);
     }
 }
